requirements.txt
absl-py==0.11.0
apipkg==1.5
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
broadbean==0.9.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
cftime==1.4.1
chardet==4.0.0
contextlib2==0.6.0.post1
coverage==5.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
execnet==1.8.0
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.12
google-api-core==1.25.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lesscpy==0.14.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mypy==0.790
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netCDF4==1.5.6
networkx==2.5
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencensus-ext-azure==1.0.6
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
ply==3.11
prettytable==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
pytz==2020.5
PyVISA==1.11.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==21.0.1
qcodes>=0.22.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
requirements-parser==0.2.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.2
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.16.2
zipp==3.4.0
